@@ -11,59 +11,59 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.4",
-    "@angular/cdk": "^16.2.3",
-    "@angular/common": "^16.2.4",
-    "@angular/compiler": "^16.2.4",
-    "@angular/core": "^16.2.4",
-    "@angular/forms": "^16.2.4",
-    "@angular/localize": "^16.2.4",
-    "@angular/platform-browser": "^16.2.4",
-    "@angular/platform-browser-dynamic": "^16.2.4",
-    "@angular/router": "^16.2.4",
-    "@angular/service-worker": "^16.2.4",
+    "@angular/animations": "^16.2.9",
+    "@angular/cdk": "^16.2.8",
+    "@angular/common": "^16.2.9",
+    "@angular/compiler": "^16.2.9",
+    "@angular/core": "^16.2.9",
+    "@angular/forms": "^16.2.9",
+    "@angular/localize": "^16.2.9",
+    "@angular/platform-browser": "^16.2.9",
+    "@angular/platform-browser-dynamic": "^16.2.9",
+    "@angular/router": "^16.2.9",
+    "@angular/service-worker": "^16.2.9",
     "@ctrl/ngx-codemirror": "^7.0.0",
     "@mean-stream/ngbx": "^0.11.0",
     "@ng-bootstrap/ng-bootstrap": "^15.1.1",
     "@popperjs/core": "^2.11.8",
-    "@sentry/angular-ivy": "^7.68.0",
+    "@sentry/angular-ivy": "^7.74.0",
     "angular-gridster2": "^16.0.0",
     "angular-split": "^16.2.1",
     "bootstrap": "~5.2.3",
     "bootstrap-darkmode": "^5.0.1",
-    "bootstrap-icons": "^1.10.5",
+    "bootstrap-icons": "^1.11.1",
     "class-validator": "^0.14.0",
     "codemirror": "^5.65.15",
     "file-saver": "^2.0.5",
-    "highlight.js": "^11.8.0",
+    "highlight.js": "^11.9.0",
     "keycloak-angular": "^14.1.0",
     "keycloak-js": "21.1.2",
-    "markdown-it": "^13.0.1",
+    "markdown-it": "^13.0.2",
     "ngx-clipboard": "^16.0.0",
-    "ngx-drag-drop": "^16.0.0",
+    "ngx-drag-drop": "^16.1.0",
     "rxjs": "~7.8.1",
     "tslib": "^2.6.2",
-    "zone.js": "^0.13.2"
+    "zone.js": "^0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.1",
-    "@angular-eslint/builder": "^16.1.2",
-    "@angular-eslint/eslint-plugin": "^16.1.2",
-    "@angular-eslint/eslint-plugin-template": "^16.1.2",
-    "@angular-eslint/schematics": "^16.1.2",
-    "@angular-eslint/template-parser": "^16.1.2",
-    "@angular/cli": "^16.2.1",
-    "@angular/compiler-cli": "^16.2.4",
-    "@angular/language-service": "^16.2.4",
+    "@angular-devkit/build-angular": "^16.2.6",
+    "@angular-eslint/builder": "^16.2.0",
+    "@angular-eslint/eslint-plugin": "^16.2.0",
+    "@angular-eslint/eslint-plugin-template": "^16.2.0",
+    "@angular-eslint/schematics": "^16.2.0",
+    "@angular-eslint/template-parser": "^16.2.0",
+    "@angular/cli": "^16.2.6",
+    "@angular/compiler-cli": "^16.2.9",
+    "@angular/language-service": "^16.2.9",
     "@types/codemirror": "^5.60.10",
     "@types/file-saver": "^2.0.5",
-    "@types/jasmine": "~4.3.5",
-    "@types/jasminewd2": "~2.0.10",
-    "@types/markdown-it": "^13.0.1",
-    "@types/node": "^20.6.0",
-    "@typescript-eslint/eslint-plugin": "^6.7.0",
-    "@typescript-eslint/parser": "^6.7.0",
-    "eslint": "^8.49.0",
+    "@types/jasmine": "~4.3.6",
+    "@types/jasminewd2": "~2.0.11",
+    "@types/markdown-it": "^13.0.2",
+    "@types/node": "^20.8.5",
+    "@typescript-eslint/eslint-plugin": "^6.7.5",
+    "@typescript-eslint/parser": "^6.7.5",
+    "eslint": "^8.51.0",
     "jasmine-core": "~5.1.1",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.4.2",
